third_party/WebKit/LayoutTests/bluetooth/requestDevice/name-empty-device-from-service-filter.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('EmptyNameHeartRateAdapter')
  .then(() => requestDeviceWithKeyDown({
    filters: [{services: ['heart_rate']}]}))
  .then(device => {
    assert_equals(device.name, '');
  })
}, 'An empty name device can be obtained by advertised service UUID.');
</script>
